Extract shared stats table logic into statsTableMixin
Refactored three stats table components (CastLineStats, CharacterLineStats,
CueCountStats) to use a shared mixin for common act/scene traversal logic.

Changes:
- Created statsTableMixin.js with sortedActs, sortedScenes computed properties
- Extracted numScenesPerAct, getHeaderName, getCellName helper methods
- Centralized act/scene data fetching in mounted lifecycle hook
- Reduced code duplication by ~150 lines across three components

Each component retains its unique stats fetching and display logic while
sharing the common act/scene hierarchy traversal pattern.
